重构

+++ app/src/main/java/widget/small/com/smallwidget/business/aui/fragment/SettingFrag.java
@@ -1,18 +1,19 @@
 package widget.small.com.smallwidget.business.aui.fragment;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.view.View;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import widget.small.com.smallwidget.R;
-import widget.small.com.smallwidget.business.aui.activity.FaceActivity;
 import widget.small.com.smallwidget.business.aui.activity.ThemeActivity;
-import widget.small.com.smallwidget.business.aui.activity.TimeActivity;
+import widget.small.com.smallwidget.business.aui.adapter.binder.FunctionItemBinder;
+import widget.small.com.smallwidget.business.bean.MainItemBean;
 import widget.small.com.smallwidget.helper.base.BaseFragment;
 import widget.small.com.smallwidget.helper.db.DaoTheme;
-import widget.small.com.smallwidget.helper.tools.CodeUtils;
 import widget.small.com.smallwidget.helper.tools.ToastUtil;
-import widget.small.com.smallwidget.helper.tools.VibrateHelp;
+import widget.small.com.smallwidget.helper.tools.base.Code;
 
 /**
  * Created by ZeroProject on 2016/5/25 17:00
@@ -25,9 +26,9 @@ import widget.small.com.smallwidget.helper.tools.VibrateHelp;
  * <p>
  * 以后SwitchLayout将会划分入我的SmartUI库下面
  */
-public class SettingFrag extends BaseFragment implements View.OnClickListener {
-    private View time, theme, clear, more;
-
+public class SettingFrag extends BaseFragment {
+    private List<MainItemBean> rvList;
+    private FunctionItemBinder binder;
 
     @Override
     protected int getLayoutId() {
@@ -36,58 +37,42 @@ public class SettingFrag extends BaseFragment implements View.OnClickListener {
 
     @Override
     protected void initView(View view) {
-        time = findView(R.id.setting_time);
-        theme = findView(R.id.setting_theme);
-        clear = findView(R.id.setting_card_3);
-        more = findView(R.id.setting_card_4);
+        rvListView = findView(R.id.rv_main_list);
+
+        rvList = new ArrayList<>();
+        binder = new FunctionItemBinder(getActivity());
+        recInit(2, MainItemBean.class, binder);
     }
 
-    @Override
-    public void onClick(View v) {
-        Intent intent = null;
-        switch (v.getId()) {
-            case R.id.setting_time://时光轴:
-                CodeUtils.showEditDialog(getActivity(), "进入时光轴需要校验你的身份", "请输入密码", new CodeUtils.PasswordCheck() {
-                    @Override
-                    public void onResult(DialogInterface dialog, String password) {
-                        if (password.equalsIgnoreCase("5277")) {
-                            dialog.dismiss();
-                            Intent intent = new Intent(context, TimeActivity.class);
-                            startActivity(intent);
-                        } else {
-                            ToastUtil.CustomShort("密码错误", ToastUtil.SHOW_DEFAULT_ICON);
-                            //10,200,10,200,10,200,10,200,400,200,10,200,10,200,10
-                            VibrateHelp.vComplicated(getActivity(), new long[]{10, 200, 400, 200, 400, 200, 10, 200, 10, 200, 400, 200, 10, 200, 10, 200, 10, 200, 10, 200, 10, 200, 10, 200, 10}, -1);
-                        }
-                    }
-                });
-                break;
-            case R.id.setting_theme://主题
-                intent = new Intent(context, ThemeActivity.class);
-                startActivity(intent);
-                break;
-            case R.id.setting_card_3://清空数据库
-                DaoTheme.getInstance().deleteAll();
-                ToastUtil.CustomShort("估计主题哪里现在应该正常了吧");
-                break;
-            case R.id.setting_card_4://
-                intent = new Intent(context, FaceActivity.class);
-                startActivity(intent);
-                break;
-        }
-    }
 
     @Override
     protected void initListener() {
-        time.setOnClickListener(this);
-        theme.setOnClickListener(this);
-        clear.setOnClickListener(this);
-        more.setOnClickListener(this);
+        binder.setEventListener(new FunctionItemBinder.EventListener() {
+            @Override
+            public void onItemClick(int position) {
+                if (position > -1 && position < rvList.size()) {
+                    int id = rvList.get(position).getId();
+                    Intent intent = null;
+                    switch (id) {
+                        case Code.StatusCode.Setting_To_Theme:
+                            intent = new Intent(context, ThemeActivity.class);
+                            startActivity(intent);
+                            break;
+                        case Code.StatusCode.Setting_to_ResetData:
+                            DaoTheme.getInstance().deleteAll();
+                            ToastUtil.CustomShort("估计主题哪里现在应该正常了吧");
+                            break;
+                    }
+                }
+            }
+        });
     }
 
     @Override
     protected void initData() {
-
+        rvList.add(new MainItemBean(R.drawable.ic_theme_g, "主题", Code.StatusCode.Setting_To_Theme));
+        rvList.add(new MainItemBean(R.drawable.ic_data_g, "重置数据", Code.StatusCode.Setting_to_ResetData));
+        recNotifyData(rvList);
     }
